+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 4.0.1
+* Fixed reported issue 1, Fixed a bug related to camera movement by user swipe.
+* Added rotationSpeed variable to Sp3dRenderer.
+
 ## 4.0.0
 * Added isAllDrawn flag to Sp3dCamera.